Extract common functionality from `internal/fileurl` into `internal/xhttp`
